refactor(spot): move domain model from pkg to internal
pkg/ holds reusable libraries (noaa, env, log, baseplate); the Spot domain
is clairBuoyant-specific app code, so it belongs in internal/. Addresses
review feedback on PR #214.

Refs #208
