+++ internal/spot/spot.go
